+++ src/signer/index.ts
@@ -1,3 +1,2 @@
 export * from './interface';
 export * from './default';
-export * from './ledger';