+++ src/types/account.ts
@@ -1,0 +1,14 @@
+import BN from 'bn.js';
+
+import { BlockIdentifier } from '../provider/utils';
+import { BigNumberish } from '../utils/number';
+import { EstimateFeeResponse } from './api';
+
+export interface EstimateFee extends EstimateFeeResponse {
+  suggestedMaxFee: BN;
+}
+
+export interface EstimateFeeDetails {
+  nonce?: BigNumberish;
+  blockIdentifier?: BlockIdentifier;
+}