+++ src/types/api.ts
@@ -1,3 +1,5 @@
+import BN from 'bn.js';
+
 import { BlockIdentifier } from '../provider/utils';
 import { BigNumberish } from '../utils/number';
 import {
@@ -43,10 +45,17 @@ export type Endpoints = {
     REQUEST: never;
     RESPONSE: GetTransactionTraceResponse;
   };
+  get_transaction_receipt: {
+    QUERY: {
+      transactionHash: string;
+    };
+    REQUEST: never;
+    RESPONSE: TransactionReceiptResponse;
+  };
   get_storage_at: {
     QUERY: {
       contractAddress: string;
-      key: number;
+      key: BigNumberish;
       blockIdentifier: BlockIdentifier;
     };
     REQUEST: never;
@@ -86,6 +95,14 @@ export type Endpoints = {
 export type GetContractAddressesResponse = {
   Starknet: string;
   GpsStatementVerifier: string;
+};
+
+export type DeclareTransaction = {
+  type: 'DECLARE';
+  contract_class: CompressedCompiledContract;
+  nonce: BigNumberish;
+  sender_address: BigNumberish;
+  signature: Signature;
 };
 
 export type DeployTransaction = {
@@ -129,7 +146,7 @@ export type ExecutionResources = {
     bitwise_builtin: number;
     output_builtin: number;
     ecdsa_builtin: number;
-    ec_op_builtin: number;
+    ec_op_builtin?: number;
   };
   n_memory_holes: number;
 };
@@ -139,7 +156,7 @@ export type CallContractTransaction = Omit<
   'type' | 'entry_point_type' | 'nonce'
 >;
 
-export type Transaction = DeployTransaction | InvokeFunctionTransaction;
+export type Transaction = DeclareTransaction | DeployTransaction | InvokeFunctionTransaction;
 
 export type CallContractResponse = {
   result: string[];
@@ -167,8 +184,9 @@ export type GetBlockResponse = {
       transaction_index: number;
     };
   };
-  previous_block_hash: string;
+  parent_block_hash: string;
   status: Status;
+  gas_price: string;
 };
 
 export type GetCodeResponse = {
@@ -178,9 +196,8 @@ export type GetCodeResponse = {
 
 export type GetTransactionStatusResponse = {
   tx_status: Status;
-  block_hash: string;
+  block_hash?: string;
   tx_failure_reason?: {
-    tx_id: number;
     code: string;
     error_message: string;
   };
@@ -194,7 +211,7 @@ export type GetTransactionTraceResponse = {
     selector: string;
     calldata: RawArgs;
     result: Array<any>;
-    execution_resources: any;
+    execution_resources: ExecutionResources;
     internal_call: Array<any>;
     events: Array<any>;
     messages: Array<any>;
@@ -202,22 +219,33 @@ export type GetTransactionTraceResponse = {
   signature: Signature;
 };
 
-export type GetTransactionResponse = {
+export type SuccessfulTransactionResponse = {
   status: Status;
   transaction: Transaction;
   block_hash: string;
   block_number: BlockNumber;
   transaction_index: number;
-  transaction_hash: string;
 };
+
+export type FailedTransactionResponse = {
+  status: 'REJECTED';
+  transaction_failure_reason: {
+    code: string;
+    error_message: string;
+  };
+  transaction: Transaction;
+};
+
+export type GetTransactionResponse = SuccessfulTransactionResponse | FailedTransactionResponse;
 
 export type AddTransactionResponse = {
   code: TransactionStatus;
   transaction_hash: string;
   address?: string;
+  class_hash?: string;
 };
 
-export type TransactionReceipt = {
+export type SuccessfulTransactionReceiptResponse = {
   status: Status;
   transaction_hash: string;
   transaction_index: number;
@@ -225,10 +253,27 @@ export type TransactionReceipt = {
   block_number: BlockNumber;
   l2_to_l1_messages: string[];
   events: string[];
+  actual_fee: string;
+  execution_resources: ExecutionResources;
 };
 
+export type FailedTransactionReceiptResponse = {
+  status: 'REJECTED';
+  transaction_failure_reason: {
+    code: string;
+    error_message: string;
+  };
+  transaction_hash: string;
+  l2_to_l1_messages: string[];
+  events: string[];
+};
+
+export type TransactionReceiptResponse =
+  | SuccessfulTransactionReceiptResponse
+  | FailedTransactionReceiptResponse;
+
 export type EstimateFeeResponse = {
-  amount: number;
+  amount: BN;
   unit: string;
 };