+++ www/docusaurus.config.js
@@ -73,7 +73,7 @@ const config = {
           },
           // {to: '/blog', label: 'Blog', position: 'left'},
           {
-            href: 'https://github.com/facebook/docusaurus',
+            href: 'https://github.com/0xs34n/starknet.js',
             label: 'GitHub',
             position: 'right',
           },